+++ src/include/cdb/cdbutil.h
@@ -212,6 +212,8 @@ extern int getgpsegmentCount(void);
 
 extern bool IsOnConflictUpdate(PlannedStmt *ps);
 
+extern void AvoidCorefileGeneration(void);
+
 #define ELOG_DISPATCHER_DEBUG(...) do { \
        if (gp_log_gang >= GPVARS_VERBOSITY_DEBUG) elog(LOG, __VA_ARGS__); \
     } while(false);
